feat: Refactor BronzeExtractor to BronzePipeline; update loading logic and SQL table definitions for consistency

+++ src/bronze/sql/create_table_pits.sql
@@ -1,10 +1,10 @@
 CREATE TABLE pits (
     date STRING,
-    driver_number INT,
+    driver_number BIGINT,
     lane_duration DOUBLE,
-    lap_number INT,
-    meeting_key INT,
-    session_key INT,
+    lap_number BIGINT,
+    meeting_key BIGINT,
+    session_key BIGINT,
     pit_duration DOUBLE,
     stop_duration DOUBLE
 )
