+++ src/bronze/sql/create_table_race_control.sql
@@ -1,12 +1,12 @@
 CREATE TABLE race_control(
     category STRING,
     date STRING,
-    driver_number INT,
-    lap_number INT,
+    driver_number BIGINT,
+    lap_number BIGINT,
     flag STRING,
-    meeting_key INT,
+    meeting_key BIGINT,
     message STRING,
-    session_key INT,
+    session_key BIGINT,
     qualifying_phase STRING,
     scope STRING,
     sector STRING
